ci-builder: add changeset
The next time that a release is triggered, build and publish
the `ci-builder` docker image. This will cause it to be used
in CI for builds, meaning that we will need to update our
foundry versions for local development.

This can be done with the following command:

```bash
foundryup -C edf15abd648bb96e2bcee342c1d72ec7d1066cd1
```