op-chain-ops: migrate error handling
The migration script entrypoint failed due to the
L1StartingBlockTag in the `DeployConfig` being set to `nil`.
This adds a `nil` check as well as some better logging.
The lower level methods for the usage of block tags are
also used instead of manually accessing the values internal
to the struct, this is what geth does internally.