contracts-bedrock: upgradable dispute game factory
Alter the dispute game factory such that it used in with
a proxy pattern. The tests are minimally modified in a way
where they now go through the proxy because the constructor
prevents usage of the contract by setting the owner to `address(0)`.
We should strive to follow this sort of pattern, although it makes
testing a bit more difficult, it forces us to test the contracts
behind proxies which is what happens in production plus it also
makes the implementation unusable such that it cannot be mistaken
as the real thing.