op-e2e: make TestBatcherMultiTx more resilient
TestBatcherMultiTx disables the batch submitter at the beginning of the
test, then starts it later to ensure that it catches up and
appropriately submits the transactions.

However, currently it does this by checking the block number before
starting the batcher, then waiting for 3 blocks before checking for the
transactions.  Because there is no synchronization, it's possible that
the block number advances before the batcher starts, which can lead to
the test flaking.

This commit changes the approach to fetching each of the next 10 blocks,
and succeeding once the transactions are found, or failing if none are
found within the next 10 blocks.