cleanup: port `replica-healthcheck` to `chain-mon`
This ports the `replica-healthcheck` service to the `chain-mon`
package. The idea is that all of the simple typescript services
that monitor the chain should live in the same package. This was
the final package that was not following this standard. This way
it will be easier to run the various monitoring services since
they will all be the same docker image, given different entrypoints.
It will also be easier to isolate and update the deps.