contracts-bedrock: foundry-ify
Move the `contracts` directory to `src` and then move
the tests into a top level `test` dir so that its
standard foundry style for a repo.

This is going to break any in flight PRs so worth trying
to get in relatively quickly, will be no fun to rebase a
bunch of PRs.

Will need to follow up with another commit that bumps the
patch versions.