poc: system config address discoverability
Solve the address discoverability by adding getters
on the system config for the L1 proxies. This is
very easy to accomplish, now the user only needs
to know the address of the systemconfig with
a particular op stack chain to know all of the
L1 contract addresses. It makes tracking all
of the chains in the superchain much easier.

There are still some todos for this PR,
a getter needs to be added for each address.