op-challenger: Fix provision of pre-state, proof and oracle data from cannon
Combines GetPreimage and GetOracleData into GetStepData. GetStepData returns the data needed to execute the step at the specified trace index, so the state is the pre-state rather than matching the claim at the specified index. The proof data contains proofs for data accessed while executing the step and oracle data the pre-images required to execute the step. This matches up correctly with the data in the cannon proof files and what is required when calling the step function.