contracts-bedrock: add deploy config validator
Adds a validator for the deploy configs so that we can be
automatically sure that deploy configs are valid. The problem
is that `getting-started.json` will never pass deploy config
validation because it is invalid JSON and it is the one that
we want to be sure is up to date. Will leave solving that to
future work because it requires a change to the getting started
guide for the op stack.