contracts-bedrock: update `.envrc.example`
We need to update the superchain getting started guide
as well as the deploy script so that the guide shows usage
of the deploy script where it only deploys a set of proxies
and then initializes them to the standard set of implementations.
For now, use `openssl` to create a random string to use as
the `create2` salt so that individual deployments do not
clash and fail. A problem was happening where contract deployments
were failing due to contract addresses colliding.