fix(ctb): Fix off-by-one in removeOwner
There was a bug in the remove all owners flow that would leave two owners on the safe.
Fixing that bug required using `swapOwners` in order to remove and
replace the last signer, since you cannot remove the only owner and then
add a new one (even when calling from a module).