contracts-bedrock: simplify safecall test
Modularizes out a bit from https://github.com/ethereum-optimism/optimism/pull/7928
where the `SafeCall` test doesn't require `CommonTest` but instead can
just use `Test`. This simplifies the `setUp` and should reduce the
amount of computation required to set up the tests. This also reduces
the need to use `vm.assume` to ignore contracts that are deployed as
part of the `CommonTest` setup.