contracts-bedrock: fuzz message passer
This commit migrates a few `L2ToL1MessagePasser` tests to being
fuzz tests instead of being regular tests. This helps to improve
coverage of the message passer and also removes unnecessary constants.
This work is done as part of making https://github.com/ethereum-optimism/optimism/pull/7928
more modular so it is easier to review.