contracts-bedrock: bump `forge-std`
Bumps `forge-std` to `37a37ab73364d6644bfe11edf88a07880f99bd56`.
This is not a release commit but it does include the latest change
to the `Vm` such that `loadAllocs(string)` is included. Pulling this
in now unblocks progress on loading in the L2 genesis state into the
solidity tests. We can update the `forge-std` version to an official
release in the future.

All diffs are updating changes to `forge-std` so that things compile
again.