contracts-bedrock: fix compile time
Reduces the compile time to about 80s on my machine. The compile time
was greatly increased as part of the migration to using the deploy
script to set up the tests. Can we have our cake and eat it to? This
commit seems to say yes.

The call to `broadcast` now needs to include `msg.sender` because the
contract doing all of the deployments is no longer the top level
contract. This assumes that the deploy contract is not nested deeper
than the 2nd callframe. The deployments would fail if that was not the
case.

The `cfg` now is behind a public getter on the deploy script, so the
majority of the commit is updating places where the deploy config is
referenced to change it into calls.