ci: use base_image constant
Previously the ubuntu base image used for the `machine` field in
ci was hardcoded. It was not standardized and sometimes different.
Updating it would require a multiline diff and was prone to missing
a spot. This commit updates the value to be a param that is used
across all of the places instead of the value being inlined.
This removes a decision to be made when adding a new job and also
allows for a 1 line diff when updating the version of the image used.