contracts-bedrock: delete old storage layout
The `.storage-layout` was replaced with a better equivalent with JSON.
The diffs in storage layouts should be easier to compare in the JSON
rather than in `.storage-layout` because of the way that the markdown
is rendered. See https://github.com/ethereum-optimism/optimism/pull/8332.

An unsolved problem is the type information is lost in the storage
layout lock now. That is something that we need to solve for in the
future.