contracts-bedrock: add base sepolia deploy config
Sources the deploy config from https://github.com/base-org/contract-deployments/blob/main/sepolia/2023-09-26-deploy/deploy-config/deploy-config.json
We need to double check that this is up to date, ie it is possible that
some config has changed over time. The problem variables that we care
about from within this config are the constructor args to the
L2OutputOracle:

- cfg.l2OutputOracleSubmissionInterval
- cfg.l2BlockTime
- cfg.l2OutputOracleProposer
- cfg.l2OutputOracleChallenger
- cfg.finalizationPeriodSeconds

If any of these values changed, they should be caught by the state diff
tests written in `superchain-ops` for performing the upgrade.

This deploy config is required to deploy a new L2OutputOracle
implementation for Base Sepolia as part of the Extended Pause Upgrade.