contracts-bedrock: fix pgn sepolia config
Updates the deploy config for pgn sepolia to
be conformant to the deploy config spec. Not sure why
CI is passing as the go code is supposed to check for
non compliant deploy config. It is likely because the solidity
and the go got out of sync. We could use a rethinking of how
we do deploy config. Ideally the entire thing can be simplified
now that deploy config contains op stack chain specific config,
superchain target specific config and also config values that
go into L1 and also config values that go into L2.

This change wouldn't be necessary if the lazy deploy config change
https://github.com/ethereum-optimism/optimism/issues/7711 was
implemented. It was attempted https://github.com/ethereum-optimism/optimism/pull/7776
but became stale and added a lot of overhead.