op-challenger: Handle Reader reading data and returning EOF in the same call (#9168)
* op-challenger: Implement verification for preimages

Identifies the leaf to challenge, but doesn't yet produce the merkle proofs required and can't actually send the transaction.

* op-challenger: Review feedback.

* op-challenger: Avoid nil

* op-challenger: Handle Reader reading data and returning EOF in the same call.

This is an unusual but valid corner case of io.Reader. Special handling is required to ensure that the state commitment is captured both before and after the last, padding only block.

* fixes

---------

Co-authored-by: refcell.eth <abigger87@gmail.com>