op-chain-ops: remove post process step (#9316)
* op-chain-ops: remove post process step

Now the the L1 genesis is generated by foundry, there should be
no need to post process the L1 genesis state. Transactions would
fail because of the resource metering values in the optimism portal
contract being set as part of the genesis setup. The setup was
originally created when Go code was used to generate the L1 genesis.

There is a small chance that this fails in CI. Will close if it fails.

* op-chain-ops: fix build

* chain-ops: delete dead argument

The `postProcess` arg is no longer used so delete it

* op-node: fix build