stateviz: remove (#9338)
Removes the stateviz code from the repo as it has not
been used in quite some time. It was removed from the
docker compose setup. This is part of cleaning up code
in the monorepo periodically to make sure that we keep
the codebase clean.

stateviz was useful for visualizing the state of the rollup.